Create MonthlyEnvelope resource
This adds the MonthlyEnvelope resource which represents the activity for
an envelope within a given month (Budget).

This also adds a new flow for creating Budget records along with the
records needed to get a budget started.

This also adds a new component for displaying the Envelope details and
its activity for the given budget (MonthlyEnvelope)