Use AGPL-3.0 as the license for this project
I spent a lot of time reading about the different licensing options that
can be used for this project.

I decided that AGPL-3.0 is the best option for Terrible because of the
following:

1. I want to keep this project open source forever
2. I want to encourage other people who want to improve the codebase to
   share their improvements by sending PRs to the main Terrible repo.
3. I don't want other people to steal the work that I've done for this
   project by forking this project, turning it into closed source, and
   profiting off of their modified version of Terrible.

I initially looked into GPL-3.0, but I was made aware that AGPL-3.0 is
an improved version of GPL-3.0 where it closes a loophole regarding the
distribution of modified versions, so here we are.

You might ask - "Why didn't you choose a more permissive license like
MIT or Apache?"

Those kinds of licenses really only make sense if you're open sourcing a
package/library. So if you make a Ruby gem, for example, you want its
license to be as permissive as possible because they will only be a part
of the bigger application that other people/companies will use.

If you slap a GPL-3.0 or AGPL-3.0 into your package, no sane person will
touch that with a 10-feet pole because using a package with that license
will require them to open source their whole project.

Terrible is not just a library; It's a complete web application. The way
I envisioned this project to be used is by hosting it officially over at
https://terrible.finance and allowing other people to self-host Terrible
on their own server/domain if they want total privacy over their data or
if they want to offer a service to other people using the official
Terrible repo.

What I don't want is for other people to launch a competing service
using a closed source and modified version of Terrible. Thus, it
makes sense that the license should be AGPL-3.0.