Associate Books and Users and major refactor (#21)
This is a huge commit and I apologize.

This initially started as an effort to associate Books and Users.
However, I realized that a lot of things need to be adjusted to
accommodate for this change.

Therefore, a major refactor also happened. I have learned so much about
the Ash Framework over the past couple of weeks, and I can now see that
there is so much room for improvement in the app.

This commit is the first step towards that.