Fix call to Ash notifiers in transactions
Some of the create functions for the resources involve creating other
records within a transaction.

Apparently, this will result in Ash notifiers being skipped, so we need
to call them explicitly to play nice.

This also adds CoreComponents to the coverage ignore as we won't be
touching this file at all. Only when Phoenix updates this module.